packages/agents/pyproject.toml
[project]
name = "workey-agents"
version = "0.1.0"
description = "Workey agent definitions, prompts, and scoring logic"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langgraph>=0.1.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
